feat: RemovableTextField 생성
